Update network layer for Medicalc [s3]
