+++ zScanner/Services/Network/Medicalc/MedicalcEndpoint.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-enum IkemEndpoint: String, Endpoint {
+enum MedicalcEndpoint: String, Endpoint {
     case documentTypes = "/documenttypes"
     case submitDocument = "/documents/summary"
     case uploadPage = "/documents/page"
