+++ zScanner/Services/Network/Medicalc/Model/AuthNetworkModel.swift
