+++ zScanner/Services/Network/Medicalc/Model/DocumentNetworkModel.swift
