+++ zScanner/Services/Network/Medicalc/Model/DocumentTypeNetworkModel.swift
