+++ zScanner/Services/Network/Medicalc/Model/FolderNetworkModel.swift
