+++ zScanner/Services/Network/Medicalc/Model/FolderSearchRequest.swift
