+++ zScanner/Services/Network/Medicalc/Model/StatusNetworkModel.swift
