+++ zScanner/Services/Network/Medicalc/Requests/SubmitPasswordRequest.swift
@@ -1,22 +1,22 @@
 //
-//  SubmitRequest.swift
+//  SubmitPasswordRequest.swift
 //  zScanner
 //
-//  Created by Jakub Skořepa on 06/08/2019.
+//  Created by Jakub Skořepa on 26/08/2019.
 //  Copyright © 2019 Institut klinické a experimentální medicíny. All rights reserved.
 //
 
 import Foundation
 
-struct SubmitReuest: Request, ParametersJsonEncoded {
+struct SubmitPasswordRequest: Request, ParametersJsonEncoded {
     typealias DataType = EmptyResponse
     
-    var endpoint: Endpoint = IkemEndpoint.submitDocument
+    var endpoint: Endpoint = MedicalcEndpoint.submitPassword
     var method: HTTPMethod = .post
     var parameters: Parameters?
     var headers: HTTPHeaders = [:]
     
-    init(document: DocumentNetworkModel) {
-        parameters = document
+    init(auth: AuthNetworkModel) {
+        parameters = auth
     }
 }
