+++ zScanner/Services/Network/Medicalc/Requests/SubmitRequest.swift
@@ -1,22 +1,22 @@
 //
-//  SubmitPasswordRequest.swift
+//  SubmitRequest.swift
 //  zScanner
 //
-//  Created by Jakub Skořepa on 26/08/2019.
+//  Created by Jakub Skořepa on 06/08/2019.
 //  Copyright © 2019 Institut klinické a experimentální medicíny. All rights reserved.
 //
 
 import Foundation
 
-struct SubmitPasswordRequest: Request, ParametersJsonEncoded {
+struct SubmitReuest: Request, ParametersJsonEncoded {
     typealias DataType = EmptyResponse
     
-    var endpoint: Endpoint = IkemEndpoint.submitPassword
+    var endpoint: Endpoint = MedicalcEndpoint.submitDocument
     var method: HTTPMethod = .post
     var parameters: Parameters?
     var headers: HTTPHeaders = [:]
     
-    init(auth: AuthNetworkModel) {
-        parameters = auth
+    init(document: DocumentNetworkModel) {
+        parameters = document
     }
 }
